SLTS-13 controller's clean-up
