+++ WebContent/WEB-INF/springDispatcher-servlet.xml
@@ -42,9 +42,6 @@
         <prop key="/logout.htm">logoutController</prop>
         <prop key="/mailing_lists.shtm">mailingListsController</prop>
         <prop key="/maintenance_events.shtm">maintenanceEventsController</prop>
-        <!-- 
-        <prop key="/point_hierarchy.shtm">pointHierarchyController</prop>
-        -->
         <prop key="/point_links.shtm">pointLinksController</prop>
         <prop key="/public_view.htm">publicViewController</prop>
         <prop key="/publisher_edit.shtm">publisherEditController</prop>
@@ -139,12 +136,6 @@
   <bean id="maintenanceEventsController" class="org.springframework.web.servlet.mvc.ParameterizableViewController">
     <property name="viewName"><value>maintenanceEvents</value></property>
   </bean>
-  
-  <!-- 
-  <bean id="pointHierarchyController" class="org.springframework.web.servlet.mvc.ParameterizableViewController">
-    <property name="viewName"><value>pointHierarchy</value></property>
-  </bean>
-  -->
 
   <bean id="pointLinksController" class="org.springframework.web.servlet.mvc.ParameterizableViewController">
     <property name="viewName"><value>pointLinks</value></property>
@@ -196,14 +187,6 @@
   <bean id="watchListController" class="com.serotonin.mango.web.mvc.controller.WatchListController">
     <property name="viewName"><value>watchList</value></property>
   </bean>
-  
-  <!-- SCADA LTS -->
-  <!-- 
-  <bean id="pointHierarchy" class="org.lts.scada.web.mvc.controller.PointHierarchyController">
-  </bean>
-  -->
-  
-  <!-- END SCADA LTS -->
   
   <bean id="webcamLiveFeedController" class="com.serotonin.mango.web.mvc.controller.WebcamLiveFeedController">
     <property name="viewName"><value>webcamLiveFeed</value></property>
